可变帧率新增 ut 用例

+++ test/unittest/hcodec_test/helper/tester_codecbase.cpp
@@ -520,9 +520,8 @@ bool TesterCodecBase::ConfigureDecoder()
         fmt.PutIntValue(MediaDescriptionKey::MD_KEY_SCALE_TYPE, static_cast<int32_t>(opt_.scaleMode.value()));
     }
     if (opt_.isVrrEnable.has_value()) {
-        fmt.PutIntValue(OHOS::Media::Tag::VIDEO_DECODER_ENABLE_VRR, static_cast<int32_t>(opt_.isVrrEnable.value()));
+        fmt.PutIntValue(OHOS::Media::Tag::VIDEO_DECODER_ENABLE_VRR, opt_.isVrrEnable.value());
     }
-
     auto begin = std::chrono::steady_clock::now();
     int32_t err = codec_->Configure(fmt);
     if (err != AVCS_ERR_OK) {
